pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "msfenicsx"
version = "0.1.0"
description = "Clean rebuild skeleton for the msfenicsx Phase 1 platform."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "PyYAML>=6.0",
    "numpy>=1.26",
    "shapely>=2.0",
    "mpi4py>=4.0",
    "petsc4py>=3.21",
    "ufl>=2024.2.0",
    "fenics-dolfinx>=0.9.0",
    "pymoo>=0.6,<0.7",
]

[project.scripts]
msfenicsx = "core.cli.main:main"
msfenicsx-evaluate = "evaluation.cli:main"
msfenicsx-optimize = "optimizers.cli:main"

[tool.setuptools.packages.find]
include = ["core*", "evaluation*", "optimizers*", "llm*", "visualization*"]

[tool.pytest.ini_options]
pythonpath = ["."]
